Remove lazy imports
Accessing a class would block the code for a few milliseconds, which is not ideal since it could block a running event loop or add unnecessary latency. While there are memory savings and overall performance to be saved with lazy imports, this isn't worth it in the grand scheme of things.